made a custom hook for app.tsx to make it cleaner
